Fixed getXmlIdsWithCorrespInOtherEditions to match xmlId and corresp properly
Before this line was used to match these parameters
corresp.value.includes(xmlId.value)
which is wrong because the match should be exact and not an inclusion.
Also the editionToSkip parameter allows the caller to avoid matching ids and corresps in the same document.
A reference check is used, using the id would be better, which the parsing doesn't set and so it's undefined.